requirements.txt
alembic==1.7.5
bcrypt==3.2.0
bidict==0.21.4
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
contextlib2==21.6.0
crontab==0.23.0
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.0
eventlet==0.33.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
freezegun==1.1.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pycparser==2.21
pyotp==2.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-engineio==4.3.1
python-socketio==5.5.1
pytz==2021.3
redis==4.1.1
requests==2.27.1
rq==1.10.1
git+https://github.com/rq/rq-scheduler.git@master#egg=rq-scheduler
schema==0.7.5
simple-websocket==0.5.0
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
validators==0.18.2
Werkzeug==2.0.2
wrapt==1.13.3
wsproto==1.0.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
